Migrate SemanticHighlighting to new expectation tests framework
