+++ test/expectations/semantic_highlighting/call_invocation.exp
@@ -1,0 +1,9 @@
+[
+  {
+    "delta_line": 0,
+    "delta_start_char": 8,
+    "length": 6,
+    "token_type": 1,
+    "token_modifiers": 0
+  }
+]
