+++ test/expectations/semantic_highlighting/call_invocation_with_variable_receiver.exp
@@ -1,0 +1,23 @@
+[
+  {
+    "delta_line": 1,
+    "delta_start_char": 2,
+    "length": 3,
+    "token_type": 0,
+    "token_modifiers": 0
+  },
+  {
+    "delta_line": 1,
+    "delta_start_char": 2,
+    "length": 3,
+    "token_type": 0,
+    "token_modifiers": 0
+  },
+  {
+    "delta_line": 0,
+    "delta_start_char": 4,
+    "length": 6,
+    "token_type": 1,
+    "token_modifiers": 0
+  }
+]
