+++ test/expectations/semantic_highlighting/command_call.exp
@@ -1,0 +1,30 @@
+[
+  {
+    "delta_line": 0,
+    "delta_start_char": 0,
+    "length": 3,
+    "token_type": 0,
+    "token_modifiers": 0
+  },
+  {
+    "delta_line": 1,
+    "delta_start_char": 0,
+    "length": 6,
+    "token_type": 1,
+    "token_modifiers": 0
+  },
+  {
+    "delta_line": 0,
+    "delta_start_char": 7,
+    "length": 10,
+    "token_type": 1,
+    "token_modifiers": 0
+  },
+  {
+    "delta_line": 0,
+    "delta_start_char": 11,
+    "length": 3,
+    "token_type": 0,
+    "token_modifiers": 0
+  }
+]
