+++ test/expectations/semantic_highlighting/command_invocation.exp
@@ -1,0 +1,9 @@
+[
+  {
+    "delta_line": 0,
+    "delta_start_char": 0,
+    "length": 4,
+    "token_type": 1,
+    "token_modifiers": 0
+  }
+]
