+++ test/expectations/semantic_highlighting/command_invocation_variable.exp
@@ -1,0 +1,23 @@
+[
+  {
+    "delta_line": 0,
+    "delta_start_char": 0,
+    "length": 3,
+    "token_type": 0,
+    "token_modifiers": 0
+  },
+  {
+    "delta_line": 1,
+    "delta_start_char": 0,
+    "length": 4,
+    "token_type": 1,
+    "token_modifiers": 0
+  },
+  {
+    "delta_line": 0,
+    "delta_start_char": 5,
+    "length": 3,
+    "token_type": 0,
+    "token_modifiers": 0
+  }
+]
