+++ test/expectations/semantic_highlighting/multi_assignment.exp
@@ -1,0 +1,30 @@
+[
+  {
+    "delta_line": 1,
+    "delta_start_char": 2,
+    "length": 1,
+    "token_type": 0,
+    "token_modifiers": 0
+  },
+  {
+    "delta_line": 0,
+    "delta_start_char": 3,
+    "length": 1,
+    "token_type": 0,
+    "token_modifiers": 0
+  },
+  {
+    "delta_line": 1,
+    "delta_start_char": 2,
+    "length": 1,
+    "token_type": 0,
+    "token_modifiers": 0
+  },
+  {
+    "delta_line": 1,
+    "delta_start_char": 2,
+    "length": 1,
+    "token_type": 0,
+    "token_modifiers": 0
+  }
+]
