+++ test/expectations/semantic_highlighting/var_aref_variable.exp
@@ -1,0 +1,9 @@
+[
+  {
+    "delta_line": 1,
+    "delta_start_char": 2,
+    "length": 1,
+    "token_type": 0,
+    "token_modifiers": 0
+  }
+]
