+++ test/expectations/semantic_highlighting/var_field_variable.exp
@@ -1,0 +1,51 @@
+[
+  {
+    "delta_line": 1,
+    "delta_start_char": 2,
+    "length": 1,
+    "token_type": 0,
+    "token_modifiers": 0
+  },
+  {
+    "delta_line": 1,
+    "delta_start_char": 2,
+    "length": 1,
+    "token_type": 0,
+    "token_modifiers": 0
+  },
+  {
+    "delta_line": 1,
+    "delta_start_char": 2,
+    "length": 1,
+    "token_type": 0,
+    "token_modifiers": 0
+  },
+  {
+    "delta_line": 1,
+    "delta_start_char": 2,
+    "length": 1,
+    "token_type": 0,
+    "token_modifiers": 0
+  },
+  {
+    "delta_line": 1,
+    "delta_start_char": 2,
+    "length": 1,
+    "token_type": 0,
+    "token_modifiers": 0
+  },
+  {
+    "delta_line": 1,
+    "delta_start_char": 2,
+    "length": 1,
+    "token_type": 0,
+    "token_modifiers": 0
+  },
+  {
+    "delta_line": 0,
+    "delta_start_char": 4,
+    "length": 1,
+    "token_type": 0,
+    "token_modifiers": 0
+  }
+]
