+++ test/expectations/semantic_highlighting/vcall_invocation.exp
@@ -1,0 +1,9 @@
+[
+  {
+    "delta_line": 1,
+    "delta_start_char": 2,
+    "length": 10,
+    "token_type": 1,
+    "token_modifiers": 0
+  }
+]
