+++ test/fixtures/aref_field.rb
@@ -1,0 +1,4 @@
+def my_method
+  a = []
+  a[1] = "foo"
+end
