+++ test/fixtures/aref_variable.rb
@@ -1,0 +1,4 @@
+def my_method
+  a = []
+  a[1]
+end
