+++ test/fixtures/call_invocation.rb
@@ -1,0 +1,1 @@
+"Hello".upcase
