+++ test/fixtures/call_invocation_with_variable_receiver.rb
@@ -1,0 +1,4 @@
+def some_method
+  var = "Hello"
+  var.upcase
+end
