+++ test/fixtures/call_invocation_with_variable_receiver_and_arguments.rb
@@ -1,0 +1,4 @@
+def some_method
+  var, arg = "Hello"
+  var.upcase(arg)
+end
