+++ test/fixtures/command_call.rb
@@ -1,0 +1,2 @@
+var = "Hello"
+object.invocation var
