+++ test/fixtures/command_invocation.rb
@@ -1,0 +1,1 @@
+puts "Hello"
