+++ test/fixtures/command_invocation_variable.rb
@@ -1,0 +1,2 @@
+var = "Hello"
+puts var
