+++ test/fixtures/fcall_invocation.rb
@@ -1,0 +1,3 @@
+def some_method
+  invocation(1, 2, 3)
+end
