+++ test/fixtures/fcall_invocation_variable_arguments.rb
@@ -1,0 +1,4 @@
+def some_method
+  var = 1
+  invocation(var)
+end
