+++ test/fixtures/local_variables.rb
@@ -1,0 +1,4 @@
+def my_method
+  var = 1
+  var
+end
