+++ test/fixtures/multi_assignment.rb
@@ -1,0 +1,5 @@
+def my_method
+  a, b = [1, 2]
+  a
+  b
+end
