+++ test/fixtures/vcall_invocation.rb
@@ -1,0 +1,3 @@
+def some_method
+  invocation
+end
